first draft project structure
